mali: platform rewrite
Samsung integrated checks on in the Pegasus platform that prevented the GPU control interfaces to work; on every decideNextStatus call the DVFS tables would be rewritten with separate fixed values. Module parameters for these values were present in the kernel so the hypothesis is that they were used for debugging during development.

-All checks are removed and the driver now looks very similar to the Orion platform.
-Removal of the CPU frequency lock to 1200MHz if the GPU is at 440MHz, this is excessive as 3D load heavy applications usually do not tax the CPU that far, and is an unnecessary power consumption burden.
-The thermal control unit temperature throttling causes to fix the voltage to a fixed value when throttling is in place; this is useless considering frequency is not limited, making the whole thing senseless. Thus removed.
-Cleaned up the files from unnecessary ifdef's for readability considering the fixed platform values.

thx to AndreiLux

ported to n80x0 by andip71